requirements.txt
plotly==5.3.1
pandas==2.2.2
streamlit== 1.35.0
pillow==10.3.0
requests==2.32.2